Created server module to export development server configuration from the package
